refactor: 템플릿 콜백 패턴 적용
